909. Snakes and Ladders
The whole big idea is using BFS,
And remember that once the node has been visited, there's no point to visit again. since the next time it's visited, it's going to be slower.
Also if the case where it mentioned that this graph can be unsolvable, think about make record and let it return -1 at the end