2130. Maximum Twin Sum of a Linked List
This solution is to first use the fast and slow pointer to get the tail and the last of the node.
After that use the slow node we have, and reverse the list starting from that part.
Then go through the list again so that we can calculate the sum with the original list and the reversed one