Better solution for 23. Merge k Sorted Lists
This is a better solution for this question. We keep sorting the list throughout every step, so that we make sure that the  first node on the top of the lists is the smallest.
Aside from that, we will also need to keep a pointer to the list that still have value in it. So that each time, we don't need to sort the whole list, but only the valid ones